requirements.txt
discord.py
yaml
random
json
requests